fix improper updates to chatMessageList and chatList
An issue was occurring where responses from the Gemini API wasn't being added to a new chat (this occurred particularly when you click on the "new chat" button). The response was instead added to the previous chat in the chatList.
Similarly, when you cold/warm boot the app, the API response to the prompt deletes the user prompt from the list, only leaving the API response.

However, after updating the logic to refresh the list collected from the database after an entry is made, the issue was fixed.